Extract bouncing button
